gpucache: unexpected NULLs due to incorrect index calculation
issue at #692. Because KDS_COLUMN_ITEM_ISNULL() internally used
uint8_t for rowid's index, so it pointed wrong bits on the nullmap.

+++ src/xpu_common.h
@@ -1360,7 +1360,7 @@ KDS_COLUMN_ITEM_ISNULL(const kern_data_store *kds,
 {
 	uint8_t	   *bitmap;
 	uint8_t		mask = (1 << (rowid & 7));
-	uint8_t		idx = (rowid >> 3);
+	uint32_t	idx = (rowid >> 3);
 
 	if (cmeta->nullmap_offset == 0)
 		return false;	/* NOT NULL */
